// FrontEnd/src/app/services/api-backend/model/departamento.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Municipio } from './municipio';


export interface Departamento { 
    idDepartamento?: number;
    nombre?: string | null;
    municipios?: Array<Municipio> | null;
}